CardInput uses custom hook to populate SRPanel
